Automate PAT
Remove all the old click on PAT button logic. Just use the automated
method to get all the screenshots for the PAT.

We just want to take a screenshot of all the results. So instead of
trying to find the exact path to show the results, we should just hack
it with plain old javascript. Then we also want to show all the text in
the results, so that's why the scroll exist. Then we want to screenshot
the more info links for the PAT.